+++ src/components/manage/Tiles/Text/Edit.jsx
@@ -20,7 +20,6 @@ import trashSVG from '../../../../icons/delete.svg';
 import addSVG from '../../../../icons/circle-plus.svg';
 import cameraSVG from '../../../../icons/camera.svg';
 import videoSVG from '../../../../icons/videocamera.svg';
-import cdSVG from '../../../../icons/cd.svg';
 import TemplatedTilesSVG from '../../../../icons/theme.svg';
 
 const messages = defineMessages({
@@ -255,15 +254,6 @@ export default class Edit extends Component {
                   })}
                 >
                   <Icon name={videoSVG} size="24px" />
-                </Button>
-                <Button
-                  icon
-                  basic
-                  onClick={this.props.onChangeTile.bind(this, this.props.tile, {
-                    '@type': 'hero',
-                  })}
-                >
-                  <Icon name={cdSVG} size="24px" />
                 </Button>
               </Button.Group>
               {tiles.customTiles.length !== 0 && (