+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changes
 
  - Fix Travis unit testing false green @sneridagh
+ - Hero Tiles @nileshgulia1
 
 ## 1.2.1 (2019-02-04)
 
@@ -48,7 +49,6 @@
 
 - Add .gitattributes file to avoid most Changelog merge conflicts @pnicolli
 - Buildout for Python 3 @pbauer
-- Hero Tiles @nileshgulia1
 - Websockets support @robgietema
 - Subrequests to search and get content actions @robgietema
 - Add logos @sneridagh @albertcasado